refactor: isolate Adobe link executor
